[#21][refactor]Card를 card 패키지로 이동
